refactor: consolidate save and create schedule functionality into ActionButton component and remove redundant SaveScheduleButton
